Centralize diffusion warmup helpers
