+++ python/sglang/multimodal_gen/runtime/managers/scheduler.py
@@ -11,7 +11,6 @@ from enum import Enum
 from typing import Any, Iterator, List
 
 import zmq
-from tqdm.auto import tqdm
 
 from sglang.multimodal_gen.runtime.disaggregation.roles import RoleType
 from sglang.multimodal_gen.runtime.disaggregation.scheduler_mixin import (
@@ -55,8 +54,7 @@ from sglang.multimodal_gen.runtime.server_args import (
     set_global_server_args,
 )
 from sglang.multimodal_gen.runtime.server_warmup import (
-    get_first_generation_req,
-    is_server_based_warmup,
+    SchedulerWarmupMixin,
     is_warmup_req,
     should_return_warmup_result,
 )
@@ -71,7 +69,7 @@ _MAX_RECV_REQS_PER_POLL = 1024
 _BATCH_METRICS_LOG_INTERVAL = 5
 
 
-class Scheduler(SchedulerPostTrainingMixin, SchedulerDisaggMixin):
+class Scheduler(SchedulerWarmupMixin, SchedulerPostTrainingMixin, SchedulerDisaggMixin):
     """
     Runs the main event loop for the rank 0 worker.
     It listens for external requests via ZMQ and coordinates with other workers.
@@ -246,103 +244,6 @@ class Scheduler(SchedulerPostTrainingMixin, SchedulerDisaggMixin):
         if len(reqs) > 1:
             return [self._dispatch_single_request(req) for req in reqs]
         return self._dispatch_single_request(reqs[0])
-
-    @staticmethod
-    def _format_warmup_req(req_or_group: Any) -> str:
-        req = get_first_generation_req(req_or_group)
-        prefix = (
-            "server warmup req"
-            if is_server_based_warmup(req_or_group)
-            else "warmup req"
-        )
-        if req is None:
-            return prefix
-
-        shape = f"{req.width}x{req.height}"
-        if req.num_frames is not None and req.num_frames > 1:
-            shape += f"x{req.num_frames}f"
-
-        default_steps = req.extra.get("cache_dit_num_inference_steps")
-        if default_steps is not None and default_steps != req.num_inference_steps:
-            steps = f"{req.num_inference_steps}/{default_steps} steps"
-        else:
-            steps = f"{req.num_inference_steps} step"
-            if req.num_inference_steps != 1:
-                steps += "s"
-
-        return f"{prefix} ({shape}, {steps})"
-
-    def _warmup_progress_total(self, req_or_group: Any | None = None) -> int:
-        req = get_first_generation_req(req_or_group)
-        if req is not None:
-            warmup_total = req.extra.get("warmup_total")
-            if warmup_total is not None:
-                return warmup_total
-
-        return max(self._warmup_total, 1)
-
-    def _ensure_warmup_progress_bar(self, req_or_group: Any) -> None:
-        if not self._show_warmup_progress:
-            return
-
-        if self._warmup_progress_bar is None:
-            self._warmup_progress_bar = tqdm(
-                total=self._warmup_progress_total(req_or_group),
-                desc="Warmup requests",
-                unit="req",
-            )
-        self._warmup_progress_bar.set_postfix_str(
-            self._format_warmup_req(req_or_group), refresh=False
-        )
-
-    def _advance_warmup_progress_bar(
-        self, req_or_group: Any, output_batch: OutputBatch
-    ) -> None:
-        if not self._show_warmup_progress:
-            return
-
-        if self._warmup_progress_bar is None:
-            self._ensure_warmup_progress_bar(req_or_group)
-
-        if output_batch.metrics is not None:
-            last_duration_s = output_batch.metrics.total_duration_s
-            self._warmup_progress_bar.set_postfix_str(
-                f"{self._format_warmup_req(req_or_group)}, last={last_duration_s:.2f}s",
-                refresh=False,
-            )
-        self._warmup_progress_bar.update(1)
-
-        if self._warmup_progress_bar.n >= self._warmup_progress_bar.total:
-            self._warmup_progress_bar.close()
-            self._warmup_progress_bar = None
-
-    def _log_warmup_result(
-        self,
-        output_batch: OutputBatch,
-        req_or_group: Any,
-        is_warmup: bool,
-    ) -> None:
-        if not is_warmup:
-            return
-
-        server_based_warmup = is_server_based_warmup(req_or_group)
-        self._warmup_processed += 1
-        self._advance_warmup_progress_bar(req_or_group, output_batch)
-
-        if output_batch.error is None:
-            if (
-                not server_based_warmup
-                and not self._logged_server_ready_after_warmup
-                and (
-                    self._warmup_total <= 0
-                    or self._warmup_processed >= self._warmup_total
-                )
-            ):
-                logger.info("The server is fired up and ready to roll!")
-                self._logged_server_ready_after_warmup = True
-        else:
-            warmup_desc = self._format_warmup_req(req_or_group)
-            logger.info(f"{warmup_desc} processing failed")
 
     def _handle_generation(
         self, reqs: list[Any], *, allow_dynamic_batching: bool = True
@@ -955,29 +856,6 @@ class Scheduler(SchedulerPostTrainingMixin, SchedulerDisaggMixin):
             stop_reason=stop_reason,
         )
         return batch_items
-
-    def process_received_reqs_with_req_based_warmup(
-        self, recv_reqs: List[tuple[bytes, Any]]
-    ) -> List[tuple[bytes, Any]]:
-        """maybe insert a cloend req before actual req to avoid compile/cold-start"""
-        if (
-            self.warmed_up
-            or not self.server_args.warmup
-            or not recv_reqs
-            or self.server_args.warmup_resolutions is not None
-            or self.server_args.server_warmup
-        ):
-            return recv_reqs
-
-        identity, req_or_group = recv_reqs[0]
-        req = get_first_generation_req(req_or_group)
-        if req is not None:
-            warmup_req = req.copy_as_warmup(self.server_args.warmup_steps)
-            recv_reqs.insert(0, (identity, warmup_req))
-            self._warmup_total = 1
-            self._warmup_processed = 0
-            self.warmed_up = True
-        return recv_reqs
 
     @staticmethod
     def _normalize_received_payload(
